Separate VolkszaehlerGateway from BaseGateway
