+++ pymeterreader/gateway/__init__.py
@@ -1,3 +1,4 @@
-from pymeterreader.gateway.gateway import BaseGateway, VolkszaehlerGateway
+from pymeterreader.gateway.basegateway import BaseGateway
+from pymeterreader.gateway.volkszaehler import VolkszaehlerGateway
 
 __all__ = ["BaseGateway", "VolkszaehlerGateway"]
